localization build capability

+++ tooling/clean.js
@@ -3,13 +3,18 @@ const fs = require('fs-extra');
 
 console.log(chalk.yellow("Good Morning, Captain"));
 
+if (!(fs.existsSync('./config/local.json'))) {
+    console.error("this workspace has not yet been set up");
+    console.error("run 'yarn setup' to begin");
+    process.exit(1);
+}
+
 return Promise.all([
     fs.readJson("./config/local.json"),
     fs.readJson("./config/project.json")
 ]).then(([localConfig, projectConfig]) => {
     return Promise.all([
-        fs.remove(`${localConfig.ttpg_folder}/${projectConfig.slug}`),
-        fs.remove(`${localConfig.ttpg_folder}/${projectConfig.slug}_dev`),
+        fs.remove(`${localConfig.ttpg_folder}/${projectConfig.variants[projectConfig.defaultVariant].slug}_dev`),
         fs.remove("./prd"),
         fs.remove("./dev"),
         fs.remove("./build")
